+++ packages/lodash-pro/scripts/build.ts
@@ -2,15 +2,16 @@ import { build } from '~scripts/build'
 import tsupConfig from './tsup.config'
 
 build({
-  dts: false,
   format: ['cjs'],
-  ...tsupConfig('cjs'),
+  ...tsupConfig,
+  dts: false,
 })
 
 build({
   dts: {
-    banner: '/// <reference types="lodash-es" />',
+    banner: '/// <reference types="lodash-es" />\n',
+    resolve: ['deepmerge'],
   },
   format: ['esm'],
-  ...tsupConfig('esm'),
+  ...tsupConfig,
 })