+++ packages/react-hook/src/index.ts
@@ -1,3 +1,4 @@
+/* eslint-disable perfectionist/sort-exports */
 /**
  * MIT License
 
@@ -21,35 +22,6 @@
  OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  SOFTWARE.
  */
-
-/**
- * This is free and unencumbered software released into the public domain.
-
- Anyone is free to copy, modify, publish, use, compile, sell, or
- distribute this software, either in source code form or as a compiled
- binary, for any purpose, commercial or non-commercial, and by any
- means.
-
- In jurisdictions that recognize copyright laws, the author or authors
- of this software dedicate any and all copyright interest in the
- software to the public domain. We make this dedication for the benefit
- of the public at large and to the detriment of our heirs and
- successors. We intend this dedication to be an overt act of
- relinquishment in perpetuity of all present and future rights to this
- software under copyright law.
-
- THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
- EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF
- MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
- IN NO EVENT SHALL THE AUTHORS BE LIABLE FOR ANY CLAIM, DAMAGES OR
- OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE,
- ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
- OTHER DEALINGS IN THE SOFTWARE.
-
- For more information, please refer to <https://unlicense.org>
- */
-
-import * as ReactUse from 'react-use'
 
 export {
   clearCache,
@@ -112,7 +84,6 @@ export {
   useSelections,
   useSessionStorageState,
   useSet,
-  useSetState,
   useSize,
   useTextSelection,
   useThrottle,
@@ -131,43 +102,6 @@ export {
   useWebSocket,
   useWhyDidYouUpdate,
 } from 'ahooks'
-
-const {
-  useAsync,
-  useAsyncFn,
-  useAsyncRetry,
-  useBeforeUnload,
-  useCopyToClipboard,
-  useCss,
-  useCustomCompareEffect,
-  useDefault,
-  useEvent,
-  useError,
-  useGetSet,
-  useGetSetState,
-  useHoverDirty,
-  useIntersection,
-  usePageLeave,
-  usePermission,
-  usePreviousDistinct,
-  useQueue,
-  useRaf,
-  useRafLoop,
-  useSearchParam,
-  useStartTyping,
-  useSpeech,
-  useSlider,
-  useScrolling,
-  useScrollbarWidth,
-  useWindowScroll,
-  useWindowSize,
-  useMeasure,
-  usePinchZoom,
-  useFirstMountState,
-  useHash,
-  useLockBodyScroll,
-} = ReactUse
-
 export {
   useAsync,
   useAsyncFn,
@@ -202,13 +136,16 @@ export {
   useStartTyping,
   useWindowScroll,
   useWindowSize,
-}
+} from 'react-use'
 
-export * from './useControlledState'
-export * from './useStrictInput'
-export * from './useEffectOnce'
-export * from './useTouch'
-export * from './useLockScroll'
-export * from './useScrollParent'
-export * from './useLayoutUpdateEffect'
-export * from './useConstant'
+export { useControlledState } from './useControlledState'
+export { useEffectOnce } from './useEffectOnce'
+export { useStrictInput } from './useStrictInput'
+export { useTouch } from './useTouch'
+export { useLockScroll } from './useLockScroll'
+export { useScrollParent } from './useScrollParent'
+export { useLayoutUpdateEffect } from './useLayoutUpdateEffect'
+export { useConstant } from './useConstant'
+export { useSetState } from './useSetState'
+export { useLocalStorageSetState } from './useLocalStorageSetState'
+export { useSessionStorageSetState } from './useSessionStorageSetState'