+++ packages/react-hook/src/types.ts
@@ -1,0 +1,3 @@
+export type RequireSome<T, K extends keyof T> = T & {
+  [P in K]-?: NonNullable<T[P]>
+}