+++ packages/react-locale/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @minko-fe/react-locale
 
+## 0.1.18
+
+### Patch Changes
+
+- bump `vite-plugin-i18n-detector`
+
+- Updated dependencies []:
+  - @minko-fe/lodash-pro@0.2.0
+  - @minko-fe/vite-config@0.1.18
+
 ## 0.1.17
 
 ### Patch Changes